Update README.md
This repository contains the Bird Strike Data Visualization project, which focuses on analysing and visualizing bird strike data collected by the FAA between 2010 and 2011. Using Python programming and libraries such as Pandas, Seaborn, Numpy, and Matplotlib, the visualizations provide valuable insights that aims to understand the patterns and impacts of bird strikes on aircraft safety.